@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.15.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,7 +26,7 @@ dependencies:
   mobx: ^2.2.1
   dio: ^5.3.3
   flutter_mobx: ^2.1.0+1
-  mobx_codegen: ^2.0.1+3
+  mobx_codegen: ^2.4.0
   mockito: ^5.0.10
   cupertino_icons: ^1.0.2
   http_mock_adapter: ^0.6.0
